Migrate package-utils from .mjs to TypeScript
